CompanyContract rework and additions

+++ EnterpriseBot.Api/Models/Settings/BusinessSettings/Company/CompanyContractSettings.cs
@@ -7,7 +7,7 @@ namespace EnterpriseBot.Api.Models.Settings.BusinessSettings.Company
 {
     public class CompanyContractSettings
     {
-        public int MaxContracts { get; set; }
-        public int MaxTimeInDays { get; set; }
+        public ushort MaxContracts { get; set; }
+        public uint MaxTimeInDays { get; set; }
     }
 }
